pyproject.toml
[project]
name = "realty-research-ai"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nehil Jain", email = "2761999+nehiljain@users.noreply.github.com"},
]
dependencies = [
    "streamlit>=1.28",
    "langchain>=0.0.217",
    "openai>=1.2",
    "duckduckgo-search",
    "anthropic>=0.3.0",
    "trubrics>=1.4.3",
    "streamlit-feedback",
    "folium>=0.16.0",
    "pandas>=2.2.1",
    "pandasai>=2.0.23",
    "streamlit-folium>=0.18.0",
    "instructor>=0.6.7",
    "pydantic>=2.6.4",
    "python-dotenv>=1.0.1",
    "tdqm>=0.0.1",
    "google-search-results>=2.4.2",
    "jupyterlab>=4.1.5",
    "ipython>=8.22.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}



[tool.pdm]
distribution = false
[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.1.1",
]
lint = [
    "ruff>=0.3.4",
]
dev = [
    "watchdog>=4.0.0",
]


